Improved button chromatic tests
