--- packages/react-components/components/button/tests/chromatic/components/index.js
@@ -1,1 +1,0 @@
-export * from "./buttons";
